Use registered batch key in `get_library_log_means_and_vars`
The `get_library_log_means_and_vars` function is hardcoded to
use a field with the name `batch` in `adata.obs`. It's not
always true that the user will name their batch key `batch`,
so this MR modifies the function to use the batches discovered
during the `setup_anndata` function.